Crate turbo_tasks_macros_shared

Structs

ValueTraitArguments

Arguments to the #[turbo_tasks::value_trait] attribute macro.

Functions

generate_destructuring

Generates a match arm destructuring pattern for the given fields.

get_function_ident

get_impl_function_ident

get_ref_ident

get_register_trait_methods_ident

get_register_value_type_ident

get_trait_default_impl_function_ident

get_trait_impl_function_ident

get_trait_type_ident

match_expansion

Handles the expansion of a struct/enum into a match statement that accesses every field for procedural code generation.